dte-service/requirements.txt
# DTE Microservice - Python Dependencies
# Python 3.11+

# ═══════════════════════════════════════════════════════════
# WEB FRAMEWORK
# ═══════════════════════════════════════════════════════════
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ═══════════════════════════════════════════════════════════
# XML PROCESSING (CRÍTICO)
# ═══════════════════════════════════════════════════════════
lxml>=4.9.0
xmlsec>=1.3.13
defusedxml>=0.7.1

# ═══════════════════════════════════════════════════════════
# FIRMA DIGITAL Y CRIPTOGRAFÍA (CRÍTICO)
# ═══════════════════════════════════════════════════════════
pyOpenSSL>=23.3.0
cryptography>=41.0.7
asn1crypto>=1.5.1

# ═══════════════════════════════════════════════════════════
# SOAP Y HTTP (CRÍTICO)
# ═══════════════════════════════════════════════════════════
zeep>=4.2.1
requests>=2.31.0
urllib3>=2.1.0
httpx>=0.25.2

# ═══════════════════════════════════════════════════════════
# ASYNC QUEUE
# ═══════════════════════════════════════════════════════════
celery>=5.3.4
redis>=5.0.1
kombu>=5.3.4

# ═══════════════════════════════════════════════════════════
# LOGGING Y MONITOREO
# ═══════════════════════════════════════════════════════════
structlog>=23.2.0
prometheus-client>=0.19.0

# ═══════════════════════════════════════════════════════════
# UTILIDADES
# ═══════════════════════════════════════════════════════════
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3

# ═══════════════════════════════════════════════════════════
# RETRY LOGIC
# ═══════════════════════════════════════════════════════════
tenacity>=8.2.3

# ═══════════════════════════════════════════════════════════
# TESTING
# ═══════════════════════════════════════════════════════════
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.25.2

